walletsite/requirements.txt
Django==4.2.23
asgiref==3.9.1
dj-database-url==2.3.0
djangorestframework==3.16.1
drf-yasg==1.21.10
gunicorn==23.0.0
whitenoise==6.7.0
packaging==25.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
pytz==2025.2
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0

